+++ src/plugins/scripts/python/weechat-python-api.c
@@ -2257,7 +2257,7 @@ weechat_python_api_hook_print_cb (void *data, struct t_gui_buffer *buffer,
     
     script_callback = (struct t_script_callback *)data;
     
-    snprintf (timebuffer, sizeof (timebuffer) - 1, "%ld", date);
+    snprintf (timebuffer, sizeof (timebuffer) - 1, "%ld", (long int)date);
     
     python_argv[0] = script_ptr2str (buffer);
     python_argv[1] = timebuffer;
